Modularize dag code by creating a util folder
• Import helper functions and boto3 calls into the gdpr-validation dag
• Create an .airflowignore so that util files aren't recognized as dags
• Delete tutorial and example dags
